refactor: move utility functions to utils folder
