+++ src/utils/index.ts
@@ -1,0 +1,6 @@
+import { ILogger } from '../types';
+import { createHttpUtils } from './http';
+
+export function createUtils(logger: ILogger) {
+  return { ...createHttpUtils(logger) };
+}
